+++ src/MagasinBundle/Controller/VendeurController.php
@@ -13,7 +13,7 @@ class VendeurController extends Controller
 
 
     public function vendeuraccAction(){
-
+        if($_SESSION['type']=="admin"){
         $em= $this->getDoctrine()->getRepository('MagasinBundle\Entity\Vendeur') ;
         $emp=$em->findAll();
         if($emp){
@@ -21,10 +21,13 @@ class VendeurController extends Controller
         }else{
             return $this->render('MagasinBundle:Default:vendeuracc.html.twig',array("error"=>"Pas de vendeur accepté!","name"=>$_SESSION['login']));
         }
+        }
+        else
+        {return $this->render('MagasinBundle:Front:error.html.twig');}
     }
 
     public function vendeurattAction(){
-
+        if($_SESSION['type']=="admin"){
         $em= $this->getDoctrine()->getRepository('MagasinBundle\Entity\Vendeur') ;
         $emp=$em->findAll();
         if($emp){
@@ -32,6 +35,9 @@ class VendeurController extends Controller
         }else{
             return $this->render('MagasinBundle:Default:vendeuratt.html.twig',array("error"=>"Pas de vendeur en attente!","name"=>$_SESSION['login']));
         }
+        }
+        else
+        {return $this->render('MagasinBundle:Front:error.html.twig');}
     }
 
 
@@ -39,7 +45,7 @@ class VendeurController extends Controller
 
     public function supprimervacAction(Vendeur $employe)
     {
-
+        if($_SESSION['type']=="admin"){
 
             $em= $this->getDoctrine()->getManager();
             $em->remove($employe);
@@ -47,27 +53,40 @@ class VendeurController extends Controller
 
         $nom=$employe->getNom();
         return $this->render('MagasinBundle:Default:supprimervac.html.twig',array("name"=>$_SESSION['login'],"nom"=>$nom));
-    }
+        }
+        else
+        {return $this->render('MagasinBundle:Front:error.html.twig');}
+        }
 
 
     public function supprimervatAction(Vendeur $employe)
     {
+
+        if($_SESSION['type']=="admin"){
         $em= $this->getDoctrine()->getManager();
         $em->remove($employe);
         $em->flush();
         $nom=$employe->getNom();
         return $this->render('MagasinBundle:Default:supprimervat.html.twig',array("name"=>$_SESSION['login'],"nom"=>$nom));
-    }
+        }
+        else
+        {return $this->render('MagasinBundle:Front:error.html.twig');}
+        }
 
     public function acceptervatAction($id)
     {
+
+        if($_SESSION['type']=="admin"){
         $em = $this->getDoctrine()->getManager();
         $employe = $em->getRepository('MagasinBundle\Entity\Vendeur')->find($id);
         $employe->setType("vendeuracc");
         $em->flush();
         $nom=$employe->getNom();
         return $this->render('MagasinBundle:Default:vendeuracc.html.twig',array("name"=>$_SESSION['login'],"nom"=>$nom));
-    }
+        }
+        else
+        {return $this->render('MagasinBundle:Front:error.html.twig');}
+        }
